add randome1

+++ Example003_Sum/Program.cs
@@ -1,5 +1,5 @@
-﻿double numberA = new Random().Next(1,10); // 1 2 3 4 ..... 9
-Console.WriteLine (numberA);
-double numberB = new Random().Next(1,10);
+﻿int numberA = new Random().Next(1,10); // 1 2 3 4 ..... 9
+Console.WriteLine(numberA);
+int numberB = new Random().Next(1,10);
 Console.WriteLine(numberB);
 Console.WriteLine(numberA + numberB);
